changed COMPILE_WITHOUT_LibRoyale to COMPILE_WITH_LibRoyale fixed a few windows bugs for LibRoyale support

+++ InfiniTAM/Engine/OpenNIEngine.cpp
@@ -192,7 +192,7 @@ OpenNIEngine::OpenNIEngine(const char *calibFilename, const char *deviceURI, con
 				(float)imageSize_d.y / (2.0f * tan(v_fov/2.0f)),
 				(float)imageSize_d.x / 2.0f,
 				(float)imageSize_d.y / 2.0f,
-				imageSize_d.x,imageSize_d.y);
+				(float)imageSize_d.x, (float)imageSize_d.y);
 		}
 		if (colorAvailable) {
 			float h_fov = data->colorStream.getHorizontalFieldOfView();
@@ -202,7 +202,7 @@ OpenNIEngine::OpenNIEngine(const char *calibFilename, const char *deviceURI, con
 				(float)imageSize_rgb.y / (2.0f * tan(v_fov/2.0f)),
 				(float)imageSize_rgb.x / 2.0f,
 				(float)imageSize_rgb.y / 2.0f,
-				imageSize_rgb.x, imageSize_rgb.y);
+				(float)imageSize_rgb.x, (float)imageSize_rgb.y);
 		}
 	}
 }
